Fix request duration logging in the JSON driver
5c6e05a changed how requests/responses were logged, and inadvertently
introduced a bug where response time info would not be logged correctly.
The implementation of the timer made such a bug easy to introduce, since
the timer state is tightly coupled with logging a message after
`end_timer` has been called.

This commit does not improve the timer implementation at all, but it
does rectify the broken assumption. The line which logs request details
should now contain both the "magic" action used and the duration of the
request. The tests have been adjusted to exercise as much, rather than
that the superclass interface was being used in a certain way.