Rescue date parsing errors in try_to_encode_as_date
A date string can fit a date or date time regular expression but be an
invalid date when parsed. For example, the "10/00/1999" will satisify
the regular expression given by `DATE_REGEX` but fail when given to
`Date.parse`. In these situations we simply rescue the error the return
the string.