checks that token is indeed a token
Previously we were counting any `200` response as having obtained a
token, but it turns out sometimes there can be an internal `404`
within the Unity instance and yet the external response is a
`200` (with a `404` HTML page as the 'token'). This checks to make
sure we are looking at a UUID in the `response.body` and not some
other content. If it's not a UUID-shaped thing, we error.