v5.0 improvements (#36)
* use a more helpful and standard message post-auth

Partly this fixes a typo, but it also uses the name of the gem and
that will make errors easier to track down.

* fail on Magic calls when user is not authenticated

* checks that token is indeed a token

Previously we were counting any `200` response as having obtained a
token, but it turns out sometimes there can be an internal `404`
within the Unity instance and yet the external response is a
`200` (with a `404` HTML page as the 'token'). This checks to make
sure we are looking at a UUID in the `response.body` and not some
other content. If it's not a UUID-shaped thing, we error.

* bump version

* removes quotes from security token fixture

* sends response body, not response, to error check

* uses Allscripts sample security token not random

Use one they tell us about in documentation, not one we randomly
generated.

* moves token regex out to constant

I also made the `"` at the start and end optional since I'm not sure
if they'll be there in the body of the real token request but _they
are_ there in our test request and this lets either option work.

* uses single quotes where appropriate

* log at the warn level when authentication fails

* collapses allow/expect where possible

In these cases you can use `expect().to recieve` instead of a double
of `allow` and then `expect().to have_recieved`.